code cleanup, added missing this-> operator and changed how storedRandomNumber vector is sized in the constructor.
Solitaire title does not print correctly.  Each letter should print one at a time with the text color changing at random for each letter.  At the end of work it then should loop back to the start and continue to print title one letter at a time with each letter a new random color.

The random colors are stored in a vector of int with 42 elements each with a random number.  42 is at the moment 3x the number of char in title vector so the title should print fully 3 times as decribed.